feat: one server bin eth + k8s aptos
