+++ docker/Dockerfile
@@ -65,7 +65,7 @@ RUN apt update && apt install -y openssl libssl-dev
 WORKDIR /app
 
 # Copy only the binaries from the builder stage
-COPY --from=builder /app/$LIGHT_CLIENT/target/release/client /app/$LIGHT_CLIENT/target/release/server_primary /app/$LIGHT_CLIENT/target/release/server_secondary ./
+COPY --from=builder /app/$LIGHT_CLIENT/target/release/client /app/$LIGHT_CLIENT/target/release/proof_server ./
 
 # Create group and user
 RUN groupadd execgroup && useradd -m -g execgroup execuser
