+++ docker/k8s/aptos-client/client-configmap.yaml
@@ -1,0 +1,6 @@
+apiVersion: v1
+kind: ConfigMap
+metadata:
+  name: aptos-node-config
+data:
+  APTOS_NODE_URL: "https://example.com"  # Replace with the actual APTOS node URL
