--- docker/k8s/client-service.yaml
@@ -1,10 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: client
-spec:
-  selector:
-    app: client
-  ports:
-    - protocol: TCP
-      port: 80
