+++ docker/k8s/proof-server/proof-server-service.yaml
@@ -1,0 +1,12 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: proof-server-service
+spec:
+  selector:
+    app: proof-server
+  ports:
+    - protocol: TCP
+      port: 80  # External port
+      targetPort: 8080  # Internal container port
+  type: ClusterIP
