--- docker/k8s/server-primary-deployment.yaml
@@ -1,30 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: server-primary
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: server-primary
-  template:
-    metadata:
-      labels:
-        app: server-primary
-    spec:
-      containers:
-        - name: server-primary
-          image: argumentcomputer/aptos-light-client:latest
-          imagePullPolicy: IfNotPresent
-          command: [ "sh", "-c", "/app/server_primary --addr ${CONTAINER_PRIMARY_ADDR}:${CONTAINER_PRIMARY_PORT} --snd-addr server-secondary:80" ]
-          env:
-            - name: RUST_LOG
-              value: "debug"
-            - name: CONTAINER_PRIMARY_ADDR
-              valueFrom:
-                fieldRef:
-                  fieldPath: status.podIP
-            - name: CONTAINER_PRIMARY_PORT
-              value: "8080"
-          ports:
-            - containerPort: 8080
