--- docker/k8s/server-primary-service.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: server-primary
-spec:
-  selector:
-    app: server-primary
-  ports:
-    - protocol: TCP
-      port: 80
-      targetPort: 8080
