--- docker/k8s/server-secondary-deployment.yaml
@@ -1,30 +1,0 @@
-apiVersion: apps/v1
-kind: Deployment
-metadata:
-  name: server-secondary
-spec:
-  replicas: 1
-  selector:
-    matchLabels:
-      app: server-secondary
-  template:
-    metadata:
-      labels:
-        app: server-secondary
-    spec:
-      containers:
-        - name: server-secondary
-          image: argumentcomputer/aptos-light-client:latest
-          imagePullPolicy: IfNotPresent
-          command: [ "sh", "-c", "/app/server_secondary --addr ${CONTAINER_SECONDARY_ADDR}:${CONTAINER_SECONDARY_PORT}" ]
-          env:
-            - name: RUST_LOG
-              value: "debug"
-            - name: CONTAINER_SECONDARY_ADDR
-              valueFrom:
-                fieldRef:
-                  fieldPath: status.podIP
-            - name: CONTAINER_SECONDARY_PORT
-              value: "8080"
-          ports:
-            - containerPort: 8080
