--- docker/k8s/server-secondary-service.yaml
@@ -1,11 +1,0 @@
-apiVersion: v1
-kind: Service
-metadata:
-  name: server-secondary
-spec:
-  selector:
-    app: server-secondary
-  ports:
-    - protocol: TCP
-      port: 80
-      targetPort: 8080
